+++ packages/pluggableWidgets/document-viewer-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- We moved the bundle of pdfjs worker to local build instead of using CDN for better CSP compliance.
+
 ## [1.0.0] - 2025-05-05
 
 ### Added